added bit when creature has no actions
